third_party/WebKit/LayoutTests/http/tests/loading/redirect-methods-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: 
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
frame "0" - didReceiveTitle: 
frame "0" - didStartProvisionalLoadForFrame
frame "0" - didCommitLoadForFrame
frame "0" - didReceiveTitle: 
frame "0" - didFinishDocumentLoadForFrame
frame "0" - didHandleOnloadEventsForFrame
frame "0" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-form.html, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method GET>
frame "0" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/loading/resources/redirect-methods-form.html, http status code 200>
frame "0" - didCommitLoadForFrame
frame "0" - didReceiveTitle: 
frame "0" - didFinishDocumentLoadForFrame
frame "0" - didHandleOnloadEventsForFrame
frame "0" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-result.php, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method POST>
frame "0" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method GET>
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true, http status code 200>
frame "0" - didCommitLoadForFrame
frame "0" - didReceiveTitle: 
frame "0" - didFinishDocumentLoadForFrame
frame "1" - didReceiveTitle: 
frame "1" - didStartProvisionalLoadForFrame
frame "1" - didCommitLoadForFrame
frame "1" - didReceiveTitle: 
frame "1" - didFinishDocumentLoadForFrame
frame "1" - didHandleOnloadEventsForFrame
frame "1" - didFinishLoadForFrame
frame "0" - didHandleOnloadEventsForFrame
frame "0" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-form.html, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method GET>
frame "1" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/loading/resources/redirect-methods-form.html, http status code 200>
frame "1" - didCommitLoadForFrame
frame "1" - didReceiveTitle: 
frame "1" - didFinishDocumentLoadForFrame
frame "1" - didHandleOnloadEventsForFrame
frame "1" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-result.php, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method POST>
frame "1" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method GET>
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true, http status code 200>
frame "1" - didCommitLoadForFrame
frame "1" - didReceiveTitle: 
frame "1" - didFinishDocumentLoadForFrame
frame "2" - didReceiveTitle: 
frame "2" - didStartProvisionalLoadForFrame
frame "2" - didCommitLoadForFrame
frame "2" - didReceiveTitle: 
frame "2" - didFinishDocumentLoadForFrame
frame "2" - didHandleOnloadEventsForFrame
frame "2" - didFinishLoadForFrame
frame "1" - didHandleOnloadEventsForFrame
frame "1" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-form.html, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method GET>
frame "2" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/loading/resources/redirect-methods-form.html, http status code 200>
frame "2" - didCommitLoadForFrame
frame "2" - didReceiveTitle: 
frame "2" - didFinishDocumentLoadForFrame
frame "2" - didHandleOnloadEventsForFrame
frame "2" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-result.php, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method POST>
frame "2" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method GET>
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true, http status code 200>
frame "2" - didCommitLoadForFrame
frame "2" - didReceiveTitle: 
frame "2" - didFinishDocumentLoadForFrame
frame "3" - didReceiveTitle: 
frame "3" - didStartProvisionalLoadForFrame
frame "3" - didCommitLoadForFrame
frame "3" - didReceiveTitle: 
frame "3" - didFinishDocumentLoadForFrame
frame "3" - didHandleOnloadEventsForFrame
frame "3" - didFinishLoadForFrame
frame "2" - didHandleOnloadEventsForFrame
frame "2" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-form.html, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method GET>
frame "3" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-form.html - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/loading/resources/redirect-methods-form.html, http status code 200>
frame "3" - didCommitLoadForFrame
frame "3" - didReceiveTitle: 
frame "3" - didFinishDocumentLoadForFrame
frame "3" - didHandleOnloadEventsForFrame
frame "3" - didFinishLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-result.php, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method POST>
frame "3" - didStartProvisionalLoadForFrame
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true, main document URL http://127.0.0.1:8000/loading/redirect-methods.html, http method POST>
http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/loading/resources/redirect-methods-result.php?redirected=true, http status code 200>
frame "3" - didCommitLoadForFrame
frame "3" - didReceiveTitle: 
frame "3" - didFinishDocumentLoadForFrame
frame "3" - didHandleOnloadEventsForFrame
frame "3" - didFinishLoadForFrame
This test checks to see what HTTP method is used to fetch the final resource in the case where the first request results in a redirect.
301, 302, 303, and 307 http redirects are all tested.
301 redirect

302 redirect

303 redirect

307 redirect
